reworked API of Transformer stack
The API of the transformer was reworked in a way that
one now gets an object through which he can access a
certain transformation.

This change allows better control over the accessed
Transformations and forces a user to pre-register the
transformations he is interested in. The pre-registration
is needed to be able to compute which "input" transformations
are needed by the transformations stack.