fixed a corner case where transform sample time is equal to the requested time
The previous code was performing a look ahead first, and then deciding
that it might not need interpolation because the requested and the
sample time are the same.
The order is now reversed, so that the transformation stream does not need
to have a lookahead sample, if both times are aligned.