when resolving producers, do not add tasks connected the input ports if one of their frames is not known
The resolution process for producers in a task T looks at all transform
producers connected to T's inputs. However, it is possible that, on
these tasks, some frames are not yet set properly.

Do not give these to the transformer resolution method, as it cannot do
anything with it (and it actually checks for that and fails)