syskit: fallback to the full task if the selected data service has no transformation output port
One thing that one definitely wants to do in the syskit transformer
configuration is use devices or definitions as e.g.

  dynamic_transformation one_dev, 'bla' => 'blo'
  dynamic_transformation one_def, 'bla' => 'blo'

The issue is that the device or definition might be a data service
(i.e. not the full task) and therefore will not have the
transformation output of the full task.

The sucky part of this change is that it is NOT guaranteed
that ANY device driver for a particular device provides a
transformation producer, which means that changing the device
driver might make the transformer resolution go boom. However,
this is such a usability issue that I prefer this corner case
to fail.