sdf: fix interpretation of joints, and move the API to an instance method on Configuration
Sorry they both end up in the same commit, but splitting it would
be really hard

This commit moves the load API to instance methods. This is so that
it can be used in the traditional transformer loading files, and that
estimators can be properly added after the SDF is loaded (it's close
to impossible, and would be really harmful, to represent them in the SDF)

It also fixes how joints are represented. Given a joint and post-joint
and pre-joint frames, SDF's convention is to:
 - make the pre-joint frame the parent link frame
 - use the joint's pose as the position of the post-joint frame within
   the child's link frame